type cast for index added
git-svn-id: https://svn.urbansim.org/src/trunk@18568 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ psrc_parcel/models/development_project_proposal_sampling_model_with_minimum.py
@@ -274,7 +274,7 @@ class DevelopmentProjectProposalSamplingModel(USDevelopmentProjectProposalSampli
 #            egligible_proposals[key] = logical_and(min_type[key] == max_type, egligible)
             utypes = unique(mean_type[wegligible])
             for value in utypes:
-                parcels_with_proposals = unique(self.proposal_set['parcel_id'][wegligible][mean_type[wegligible]==value])
+                parcels_with_proposals = (unique(self.proposal_set['parcel_id'][wegligible][mean_type[wegligible]==value])).astype(int32)
                 chosen_prop = array(maximum_position(self.weight[wegligible], 
                                         labels=(self.proposal_set['parcel_id'][wegligible])*(mean_type[wegligible]==value), 
                                         index=parcels_with_proposals)).flatten().astype(int32)
